refactor: 포트폴리오 패키지 정리 (#501, #519)
