+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/application/PortfolioService.java
@@ -17,14 +17,14 @@ import com.woowacourse.pickgit.portfolio.application.dto.response.ContactRespons
 import com.woowacourse.pickgit.portfolio.application.dto.response.PortfolioResponseDto;
 import com.woowacourse.pickgit.portfolio.application.dto.response.ProjectResponseDto;
 import com.woowacourse.pickgit.portfolio.application.dto.response.SectionResponseDto;
-import com.woowacourse.pickgit.portfolio.domain.Contact;
-import com.woowacourse.pickgit.portfolio.domain.Description;
-import com.woowacourse.pickgit.portfolio.domain.Item;
+import com.woowacourse.pickgit.portfolio.domain.contact.Contact;
+import com.woowacourse.pickgit.portfolio.domain.section.item.Description;
+import com.woowacourse.pickgit.portfolio.domain.section.item.Item;
 import com.woowacourse.pickgit.portfolio.domain.Portfolio;
-import com.woowacourse.pickgit.portfolio.domain.PortfolioRepository;
-import com.woowacourse.pickgit.portfolio.domain.Project;
-import com.woowacourse.pickgit.portfolio.domain.ProjectTag;
-import com.woowacourse.pickgit.portfolio.domain.Section;
+import com.woowacourse.pickgit.portfolio.domain.repository.PortfolioRepository;
+import com.woowacourse.pickgit.portfolio.domain.project.Project;
+import com.woowacourse.pickgit.portfolio.domain.project.ProjectTag;
+import com.woowacourse.pickgit.portfolio.domain.section.Section;
 import com.woowacourse.pickgit.tag.domain.Tag;
 import com.woowacourse.pickgit.tag.domain.TagRepository;
 import java.util.ArrayList;
