+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/application/dto/response/ContactResponseDto.java
@@ -1,6 +1,6 @@
 package com.woowacourse.pickgit.portfolio.application.dto.response;
 
-import com.woowacourse.pickgit.portfolio.domain.Contact;
+import com.woowacourse.pickgit.portfolio.domain.contact.Contact;
 import com.woowacourse.pickgit.portfolio.presentation.dto.request.ContactRequest;
 import lombok.Builder;
