+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/domain/contact/Contacts.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.portfolio.domain;
+package com.woowacourse.pickgit.portfolio.domain.contact;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
