+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/domain/project/ProjectType.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.portfolio.domain;
+package com.woowacourse.pickgit.portfolio.domain.project;
 
 import com.woowacourse.pickgit.exception.portfolio.ProjectTypeNotFoundException;
 import java.util.Arrays;
