+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/domain/project/Projects.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.portfolio.domain;
+package com.woowacourse.pickgit.portfolio.domain.project;
 
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
