+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/portfolio/domain/section/item/Description.java
@@ -1,5 +1,6 @@
-package com.woowacourse.pickgit.portfolio.domain;
+package com.woowacourse.pickgit.portfolio.domain.section.item;
 
+import com.woowacourse.pickgit.portfolio.domain.section.item.Item;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
